Added rails.relative_url_append context parameter.
This can be set in your web.xml or similar to append a string to the end of the context path when the RAILS_RELATIVE_URL_ROOT is set. This allows you to host your rails app in a location other than the base context path of your application.

Placing the application at a subpath of the context root is useful when you are serving rails pages from the same WAR as another framework. For instance, you can have a single WAR that serves both JSF and Rails pages.

If you want to do this, you'll also have to set the RackFilter path mapping to reflect the new URL for the parts of the application that should be served by Rack.