Use a ; instead of \n when evaluating rackup code
Using a \n here means the lines in the evaluated rackup code don't match up.
This is made even harder to userstand because evalScriptlet() doesn't know
the originating filename, so it's hard to even infer that errors in config.ru
match up with what this is evaluating.

This change should (in theory) make errors in config.ru match up with their
actual lines in the file, even though there's nothing to tell you the code
comes from config.ru.

I have no idea how to test this!