Log the environment loading exception raised from Factory
## Overview
when loading the environment of a jruby-rack based app, such as a RoR
app, if something fails in the loading process and if [properties
settings] is not set, then the exception is left unreported, which can
be confusing or misleading. This seems to be [assumed] in the
application tool.

[properties
settings]:https://github.com/jruby/jruby-rack/blob/d59193e9284bce30a7cae
5a8ebb7f4ae8436e110/src/main/java/org/jruby/rack/DefaultRackConfig.java#
L275-L285
[assumed]:https://github.com/jruby/jruby-rack/blob/d59193e9284bce30a7cae
5a8ebb7f4ae8436e110/src/main/java/org/jruby/rack/RackServletContextListe
ner.java#L96

## Changes

1. add the logic to catch any RaiseException, output the error
stacktrace to error log, and raise a new `RackInitializationException`.