[test] Drop testing against Rails < 6.1
The logger changes on master right now only actually work on 6.0 onwards. 6.0 has been EOL for 27 months now so seems reasonable to remove. 6.1 has been EOL for 11 months, but given the rails EOL policy is pretty aggressive and there are not many special changes to handle Rails 6.1 vs 7.0 it seems reasonable to keep it for now.