[chore] Remove unnecessary deprecated RackLogger override signatures
These turn up as warnings for all usages within juby-rack itself, since most usages use a string. The interface was changed to accept a more generic type, so this is only a breaking change for sub-classes rather than consumers. Since 1.3 has other changes to the logging implementations, we may as well make this required too.