[fix] Fix JSP forward/inclusion by monkey-patching the "correct" class
This logic has been moved to re-usable helpers a long time back, which Rails and other frameworks rely upon. Patching modules already included doesn't appear to work on Ruby 2.6/JRuby 9.3 so add a fallback to patch the classes known to include. Don't know a better way; but this logic will be dropped on jruby-rack 1.3.